requirements.txt
amqp==2.6.0
APScheduler==3.6.3
asgiref==3.2.7
autopep8==1.4.3
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
Brotli==1.0.7
celery==4.4.6
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
click-log==0.3.2
cloudinary==1.21.1
cloudinary-cli==1.0.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-allauth==0.42.0
django-appconf==1.0.4
django-autoslug==1.9.7
django-background-tasks==1.2.5
django-ckeditor==5.9.0
django-cloudinary-storage==0.2.3
django-compat==1.0.15
django-compressor==2.2
django-cors-headers==3.4.0
django-el-pagination==3.1.0
django-environ==0.4.5
django-heroku==0.3.1
django-js-asset==1.2.2
django-libsass==0.7
django-modelcluster==4.4.1
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-smtp-ssl==1.0
django-taggit==1.3.0
django-tinymce4-lite==1.8.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.2
djangorestframework==3.11.0
draftjs-exporter==2.0.0
flake8==3.6.0
future==0.18.2
gunicorn==20.0.4
html5lib==0.999999999
idna==2.9
importlib-metadata==1.7.0
isort==4.3.4
Jinja2==2.11.2
jsmin==2.2.2
kombu==4.6.11
libsass==0.20.0
Markdown==2.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
olefile==0.46
pendulum==2.1.0
phonenumbers==8.12.5
Pillow==4.3.0
psycopg2==2.8.5
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.6.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
rcssmin==1.0.6
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.0.12
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tzlocal==2.1
Unidecode==0.4.21
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
whitenoise==5.1.0
Willow==1.1
zipp==3.1.0
